Put game instructions into games.py
